+++ src/backend/d3d12/BindGroupLayoutD3D12.cpp
@@ -14,7 +14,7 @@
 
 #include "backend/d3d12/BindGroupLayoutD3D12.h"
 
-#include "backend/d3d12/D3D12Backend.h"
+#include "backend/d3d12/DeviceD3D12.h"
 #include "common/BitSetIterator.h"
 
 namespace backend { namespace d3d12 {
