+++ src/backend/d3d12/GeneratedCodeIncludes.h
@@ -18,8 +18,8 @@
 #include "backend/d3d12/BufferD3D12.h"
 #include "backend/d3d12/CommandBufferD3D12.h"
 #include "backend/d3d12/ComputePipelineD3D12.h"
-#include "backend/d3d12/D3D12Backend.h"
 #include "backend/d3d12/DepthStencilStateD3D12.h"
+#include "backend/d3d12/DeviceD3D12.h"
 #include "backend/d3d12/InputStateD3D12.h"
 #include "backend/d3d12/PipelineLayoutD3D12.h"
 #include "backend/d3d12/QueueD3D12.h"
