+++ src/backend/d3d12/RenderPipelineD3D12.cpp
@@ -15,8 +15,8 @@
 #include "backend/d3d12/RenderPipelineD3D12.h"
 
 #include "backend/d3d12/BlendStateD3D12.h"
-#include "backend/d3d12/D3D12Backend.h"
 #include "backend/d3d12/DepthStencilStateD3D12.h"
+#include "backend/d3d12/DeviceD3D12.h"
 #include "backend/d3d12/InputStateD3D12.h"
 #include "backend/d3d12/PipelineLayoutD3D12.h"
 #include "backend/d3d12/ShaderModuleD3D12.h"
